+++ rabix-bindings-sb/src/main/java/org/rabix/bindings/sb/helper/SBSchemaHelper.java
@@ -383,13 +383,13 @@ public class SBSchemaHelper extends SBBeanHelper {
       return new DataType(DataType.Type.INT, !isRequired(schema));
     }
     if (isTypeFromSchema(schema, "long")) {
-      return new DataType(DataType.Type.LONG, !isRequired(schema));
+      return new DataType(DataType.Type.INT, !isRequired(schema));
     }
     if (isTypeFromSchema(schema, "float")) {
       return new DataType(DataType.Type.FLOAT, !isRequired(schema));
     }
     if (isTypeFromSchema(schema, "double")) {
-      return new DataType(DataType.Type.DOUBLE, !isRequired(schema));
+      return new DataType(DataType.Type.FLOAT, !isRequired(schema));
     }
     if (isTypeFromSchema(schema, SCHEMA_NULL)) {
       return new DataType(DataType.Type.NULL);
